refactor: api 폴더 구조 수정
- index 파일 제거
