+++ src/apis/job.ts
@@ -1,5 +1,5 @@
 import { APIAllJob } from '@/types/job';
-import { publicApi } from '../core/axios';
+import { publicApi } from './core/axios';
 
 const jobAPI = {
   getAllJobs: () => publicApi.get<APIAllJob>(`/service-api/jobs`),
