+++ src/queries/recommend/useAuthRecommendedBooks/index.tsx
@@ -1,12 +1,12 @@
-import RecommendAPI from '@/apis/recommend';
+import recommendAPI from '@/apis/recommend';
 import { APIJobGroup } from '@/types/job';
 import { useQuery } from '@tanstack/react-query';
 
 const useAuthRecommendedBooks = (jobGroup: APIJobGroup['name']) =>
   useQuery(['authRecommendedBooks', jobGroup], () =>
-    RecommendAPI.getAuthRecommendedBooks(jobGroup).then(
-      response => response.data
-    )
+    recommendAPI
+      .getAuthRecommendedBooks(jobGroup)
+      .then(response => response.data)
   );
 
 export default useAuthRecommendedBooks;
