+++ src/queries/recommend/useAuthRecommendedBookshelf/index.tsx
@@ -1,12 +1,12 @@
-import RecommendAPI from '@/apis/recommend';
+import recommendAPI from '@/apis/recommend';
 import { APIJobGroup } from '@/types/job';
 import { useQuery } from '@tanstack/react-query';
 
 const useAuthRecommendedBookshelf = (jobGroup: APIJobGroup['name']) =>
   useQuery(['authRecommendedBookshelf', jobGroup], () =>
-    RecommendAPI.getAuthRecommendedBookshelf(jobGroup).then(
-      response => response.data
-    )
+    recommendAPI
+      .getAuthRecommendedBookshelf(jobGroup)
+      .then(response => response.data)
   );
 
 export default useAuthRecommendedBookshelf;
