+++ src/queries/recommend/useUnAuthRecommendedBookshelfQuery/index.tsx
@@ -1,9 +1,9 @@
-import RecommendAPI from '@/apis/recommend';
+import recommendAPI from '@/apis/recommend';
 import { useQuery } from '@tanstack/react-query';
 
 const useUnAuthRecommendedBookshelfQuery = () =>
   useQuery(['unAuthRecommendedBookshelf'], () =>
-    RecommendAPI.getUnAuthRecommendedBookshelf().then(response => response.data)
+    recommendAPI.getUnAuthRecommendedBookshelf().then(response => response.data)
   );
 
 export default useUnAuthRecommendedBookshelfQuery;
